ScaffNet_project/requirements.txt
# ScaffNet项目依赖
# 基础科学计算库
numpy>=1.19.0
scipy>=1.6.0

# 深度学习框架
torch>=1.8.0
torchvision>=0.9.0

# 配置文件处理
pyyaml>=5.4.0

# 数据处理与可视化
pandas>=1.2.0
matplotlib>=3.3.0

# 其他工具
tqdm>=4.50.0  # 进度条
scikit-learn>=0.24.0  # 用于评估指标 